Trying REAL data type

+++ backend/db/migrations/20230220224141-create-spot.js
@@ -37,11 +37,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.NUMERIC,
+        type: Sequelize.REAL,
         allowNull: false
       },
       lng: {
-        type: Sequelize.NUMERIC,
+        type: Sequelize.REAL,
         allowNull: false
       },
       name: {
@@ -53,7 +53,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.NUMERIC,
+        type: Sequelize.REAL,
         allowNull: false
       },
       createdAt: {
